[YW2-10] feat: module setting
