+++ src/app/api/build.gradle.kts
@@ -15,6 +15,9 @@ dependencies {
 
     implementation(project(":auth"))
     implementation(project(":volunteer"))
+    implementation(project(":volunteerEvent"))
+    implementation(project(":shelter"))
+    implementation(project(":observationAnimal"))
 
     implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("org.springframework.boot:spring-boot-starter-actuator")
