+++ src/domain/shelter/build.gradle.kts
@@ -1,0 +1,16 @@
+import org.springframework.boot.gradle.tasks.bundling.BootJar
+plugins {
+    kotlin("jvm")
+}
+
+dependencies {
+    compileOnly("org.springframework:spring-tx")
+    compileOnly("org.springframework:spring-context")
+}
+tasks.named<BootJar>("bootJar") {
+    enabled = false
+}
+
+tasks.named<Jar>("jar") {
+    enabled = true
+}
