+++ src/infrastructure/storage/src/main/kotlin/yapp/be/storage/jpa/repository/ObservationAnimalTagJpaRepository.kt
@@ -1,4 +1,4 @@
-package yapp.be.storage.jpa
+package yapp.be.storage.jpa.repository
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
