+++ src/infrastructure/storage/src/main/kotlin/yapp/be/storage/repository/ObservationAnimalRepository.kt
@@ -1,12 +1,12 @@
-package yapp.be.storage
+package yapp.be.storage.repository
 
 import org.springframework.stereotype.Component
 import yapp.be.domain.port.outbound.VolunteerQueryHandler
-import yapp.be.storage.jpa.VolunteerWaitingQueueJpaRepository
+import yapp.be.storage.jpa.repository.ObservationAnimalJpaRepository
 
 @Component
-class VolunteerWaitingQueueRepository(
-    private val jpaRepository: VolunteerWaitingQueueJpaRepository
+class ObservationAnimalRepository(
+    private val jpaRepository: ObservationAnimalJpaRepository
 ) : VolunteerQueryHandler {
     override fun countAll(): Int {
         return jpaRepository.count().toInt()
