+++ src/infrastructure/storage/src/main/kotlin/yapp/be/storage/repository/ObservationAnimalTagMappingRepository.kt
@@ -1,12 +1,12 @@
-package yapp.be.storage
+package yapp.be.storage.repository
 
 import org.springframework.stereotype.Component
 import yapp.be.domain.port.outbound.VolunteerQueryHandler
-import yapp.be.storage.jpa.ObservationAnimalJpaRepository
+import yapp.be.storage.jpa.repository.ObservationAnimalTagMappingJpaRepository
 
 @Component
-class ObservationAnimalRepository(
-    private val jpaRepository: ObservationAnimalJpaRepository
+class ObservationAnimalTagMappingRepository(
+    private val jpaRepository: ObservationAnimalTagMappingJpaRepository
 ) : VolunteerQueryHandler {
     override fun countAll(): Int {
         return jpaRepository.count().toInt()
