+++ src/infrastructure/storage/src/main/kotlin/yapp/be/storage/repository/ObservationAnimalTagRepository.kt
@@ -1,12 +1,12 @@
-package yapp.be.storage
+package yapp.be.storage.repository
 
 import org.springframework.stereotype.Component
 import yapp.be.domain.port.outbound.VolunteerQueryHandler
-import yapp.be.storage.jpa.VolunteerActivityTypeJpaRepository
+import yapp.be.storage.jpa.repository.ObservationAnimalTagJpaRepository
 
 @Component
-class VolunteerActivityTypeRepository(
-    private val jpaRepository: VolunteerActivityTypeJpaRepository
+class ObservationAnimalTagRepository(
+    private val jpaRepository: ObservationAnimalTagJpaRepository
 ) : VolunteerQueryHandler {
     override fun countAll(): Int {
         return jpaRepository.count().toInt()
