+++ src/infrastructure/storage/src/main/kotlin/yapp/be/storage/repository/ShelterRepository.kt
@@ -1,12 +1,12 @@
-package yapp.be.storage
+package yapp.be.storage.repository
 
 import org.springframework.stereotype.Component
 import yapp.be.domain.port.outbound.VolunteerQueryHandler
-import yapp.be.storage.jpa.VolunteerJpaRepository
+import yapp.be.storage.jpa.repository.ShelterJpaRepository
 
 @Component
-class VolunteerRepository(
-    private val jpaRepository: VolunteerJpaRepository
+class ShelterRepository(
+    private val jpaRepository: ShelterJpaRepository
 ) : VolunteerQueryHandler {
     override fun countAll(): Int {
         return jpaRepository.count().toInt()
