+++ src/infrastructure/storage/src/main/kotlin/yapp/be/storage/repository/VolunteerEventMappingRepository.kt
@@ -1,12 +1,12 @@
-package yapp.be.storage
+package yapp.be.storage.repository
 
 import org.springframework.stereotype.Component
 import yapp.be.domain.port.outbound.VolunteerQueryHandler
-import yapp.be.storage.jpa.ShelterAdminJpaRepository
+import yapp.be.storage.jpa.repository.VolunteerEventMappingJpaRepository
 
 @Component
-class ShelterAdminRepository(
-    private val jpaRepository: ShelterAdminJpaRepository
+class VolunteerEventMappingRepository(
+    private val jpaRepository: VolunteerEventMappingJpaRepository
 ) : VolunteerQueryHandler {
     override fun countAll(): Int {
         return jpaRepository.count().toInt()
