+++ src/infrastructure/storage/src/main/kotlin/yapp/be/storage/repository/VolunteerEventRepository.kt
@@ -1,8 +1,8 @@
-package yapp.be.storage
+package yapp.be.storage.repository
 
 import org.springframework.stereotype.Component
 import yapp.be.domain.port.outbound.VolunteerQueryHandler
-import yapp.be.storage.jpa.VolunteerEventJpaRepository
+import yapp.be.storage.jpa.repository.VolunteerEventJpaRepository
 
 @Component
 class VolunteerEventRepository(
