+++ src/infrastructure/storage/src/main/kotlin/yapp/be/storage/repository/VolunteerJoinQueueRepository.kt
@@ -1,12 +1,12 @@
-package yapp.be.storage
+package yapp.be.storage.repository
 
 import org.springframework.stereotype.Component
 import yapp.be.domain.port.outbound.VolunteerQueryHandler
-import yapp.be.storage.jpa.ObservationAnimalTagJpaRepository
+import yapp.be.storage.jpa.repository.VolunteerJoinQueueJpaRepository
 
 @Component
-class ObservationAnimalTagRepository(
-    private val jpaRepository: ObservationAnimalTagJpaRepository
+class VolunteerJoinQueueRepository(
+    private val jpaRepository: VolunteerJoinQueueJpaRepository
 ) : VolunteerQueryHandler {
     override fun countAll(): Int {
         return jpaRepository.count().toInt()
