+++ src/infrastructure/storage/src/main/kotlin/yapp/be/storage/repository/VolunteerRepository.kt
@@ -1,12 +1,12 @@
-package yapp.be.storage
+package yapp.be.storage.repository
 
 import org.springframework.stereotype.Component
 import yapp.be.domain.port.outbound.VolunteerQueryHandler
-import yapp.be.storage.jpa.VolunteerEventActivityTypeMappingJpaRepository
+import yapp.be.storage.jpa.repository.VolunteerJpaRepository
 
 @Component
-class VolunteerEventActivityTypeMappingRepository(
-    private val jpaRepository: VolunteerEventActivityTypeMappingJpaRepository
+class VolunteerRepository(
+    private val jpaRepository: VolunteerJpaRepository
 ) : VolunteerQueryHandler {
     override fun countAll(): Int {
         return jpaRepository.count().toInt()
