+++ src/infrastructure/storage/src/main/kotlin/yapp/be/storage/repository/VolunteerWaitingQueueRepository.kt
@@ -1,12 +1,12 @@
-package yapp.be.storage
+package yapp.be.storage.repository
 
 import org.springframework.stereotype.Component
 import yapp.be.domain.port.outbound.VolunteerQueryHandler
-import yapp.be.storage.jpa.ShelterJpaRepository
+import yapp.be.storage.jpa.repository.VolunteerWaitingQueueJpaRepository
 
 @Component
-class ShelterRepository(
-    private val jpaRepository: ShelterJpaRepository
+class VolunteerWaitingQueueRepository(
+    private val jpaRepository: VolunteerWaitingQueueJpaRepository
 ) : VolunteerQueryHandler {
     override fun countAll(): Int {
         return jpaRepository.count().toInt()
